Updated specs for reports controller
