+++ spec/requests/api/v1/reports_calibration_spec.rb
@@ -48,6 +48,51 @@ RSpec.describe 'Reports calibration', type: :request do
     documentation
   end
 
+  # ---------------------------------------------------------------------------
+  # Calibration participant management (add / list / remove)
+  # ---------------------------------------------------------------------------
+  describe 'POST /assignments/:assignment_id/review_mappings/calibration_participants' do
+    let(:submitter_user) do
+      User.create!(name: 'unctlt1', full_name: 'UNC TLT 1',
+                   email: 'unctlt1@example.com', password: 'password',
+                   role: student_role, institution: institution)
+    end
+    let(:participants_path) { "/assignments/#{assignment.id}/review_mappings/calibration_participants" }
+
+    it 'atomically creates participant, team, and for_calibration map and returns 201' do
+      submitter_user
+      expect {
+        post participants_path, params: { username: 'unctlt1' }, headers: instructor_headers
+      }.to change { AssignmentParticipant.where(parent_id: assignment.id, user_id: submitter_user.id).count }.by(1)
+        .and change { AssignmentTeam.where(parent_id: assignment.id).count }.by(1)
+        .and change { ReviewResponseMap.where(reviewed_object_id: assignment.id, for_calibration: true).count }.by(1)
+
+      expect(response).to have_http_status(:created)
+      json = JSON.parse(response.body)
+      expect(json).to include('username' => 'unctlt1', 'instructor_review_status' => 'not_started')
+    end
+
+    it 'returns 404 when the username does not exist' do
+      post participants_path, params: { username: 'nobody' }, headers: instructor_headers
+      expect(response).to have_http_status(:not_found)
+    end
+
+    it 'is idempotent — re-adding the same username does not create duplicate maps or teams' do
+      submitter_user
+      post participants_path, params: { username: 'unctlt1' }, headers: instructor_headers
+      expect(response).to have_http_status(:created)
+
+      maps_before  = ReviewResponseMap.where(reviewed_object_id: assignment.id, for_calibration: true).count
+      teams_before = AssignmentTeam.where(parent_id: assignment.id).count
+
+      post participants_path, params: { username: 'unctlt1' }, headers: instructor_headers
+
+      expect(response).to have_http_status(:created)
+      expect(ReviewResponseMap.where(reviewed_object_id: assignment.id, for_calibration: true).count).to eq(maps_before)
+      expect(AssignmentTeam.where(parent_id: assignment.id).count).to eq(teams_before)
+    end
+  end
+
   describe 'GET /assignments/:assignment_id/reports/calibration/:map_id' do
     it 'returns report JSON for a calibration map' do
       create_response(
@@ -236,5 +281,75 @@ RSpec.describe 'Reports calibration', type: :request do
     end
 
     response
+  end
+end
+
+# ---------------------------------------------------------------------------
+# Service unit tests for Reports::CalibrationReport (bucket logic in isolation)
+# ---------------------------------------------------------------------------
+RSpec.describe Reports::CalibrationReport do
+  let(:assignment)             { create(:assignment) }
+  let(:reviewee_team)          { create(:assignment_team, assignment: assignment) }
+  let(:instructor_participant) { create(:assignment_participant, assignment: assignment) }
+  let(:instructor_map) do
+    create(:review_response_map, :for_calibration,
+           assignment: assignment, reviewer: instructor_participant, reviewee: reviewee_team)
+  end
+  let(:questionnaire) do
+    Questionnaire.create!(
+      name: "Calibration Rubric #{SecureRandom.hex(4)}",
+      private: false, min_question_score: 0, max_question_score: 5,
+      instructor: Instructor.find(instructor_participant.user_id)
+    )
+  end
+  let!(:code_quality) { create(:item, questionnaire: questionnaire, txt: 'Code quality', seq: 1) }
+  let!(:documentation) { create(:item, questionnaire: questionnaire, txt: 'Documentation', seq: 2) }
+
+  before { AssignmentQuestionnaire.create!(assignment: assignment, questionnaire: questionnaire, used_in_round: 1) }
+
+  def make_response(map:, submitted:, scores:, updated_at: Time.current)
+    r = Response.create!(response_map: map, round: 1, version_num: 1,
+                         is_submitted: submitted, created_at: updated_at, updated_at: updated_at)
+    scores.each { |item, score| Answer.create!(response: r, item: item, answer: score[:answer], comments: score[:comments]) }
+    r
+  end
+
+  it 'accumulates bucket counts from submitted student responses' do
+    make_response(map: instructor_map, submitted: true,
+                  scores: { code_quality => { answer: 4 }, documentation => { answer: 5 } })
+    student_map = create(:review_response_map, :for_calibration, assignment: assignment,
+                         reviewer: create(:assignment_participant, assignment: assignment), reviewee: reviewee_team)
+    make_response(map: student_map, submitted: true,
+                  scores: { code_quality => { answer: 3, comments: 'ok' }, documentation => { answer: 5 } })
+
+    result = described_class.new(instructor_map).render
+    cq = result[:per_item_summary].find { |s| s[:item_id] == code_quality.id }
+    expect(cq[:instructor_score]).to eq(4)
+    expect(cq[:bucket_counts]['3']).to eq(1)
+    expect(cq[:bucket_counts]['4']).to eq(0)
+  end
+
+  it 'uses only the latest submitted response per student map' do
+    make_response(map: instructor_map, submitted: true,
+                  scores: { code_quality => { answer: 4 }, documentation => { answer: 5 } })
+    student_map = create(:review_response_map, :for_calibration, assignment: assignment,
+                         reviewer: create(:assignment_participant, assignment: assignment), reviewee: reviewee_team)
+    make_response(map: student_map, submitted: true, updated_at: 2.days.ago,
+                  scores: { code_quality => { answer: 1 }, documentation => { answer: 2 } })
+    make_response(map: student_map, submitted: true, updated_at: 1.day.ago,
+                  scores: { code_quality => { answer: 3 }, documentation => { answer: 2 } })
+    make_response(map: student_map, submitted: false, updated_at: 1.hour.ago,
+                  scores: { code_quality => { answer: 5 }, documentation => { answer: 5 } })
+
+    result = described_class.new(instructor_map).render
+    cq = result[:per_item_summary].find { |s| s[:item_id] == code_quality.id }
+    expect(cq[:bucket_counts]['3']).to eq(1)
+    expect(cq[:bucket_counts]['1']).to eq(0)
+    expect(cq[:bucket_counts]['5']).to eq(0)
+  end
+
+  it 'raises InstructorResponseMissing when the instructor has not submitted' do
+    expect { described_class.new(instructor_map).render }
+      .to raise_error(Reports::CalibrationReport::InstructorResponseMissing)
   end
 end
